Many changes. Including a template for display interfaces. A seperate window for operators (so that when you click in that window, the operation takes place; TouchData is a method of generalizing touch events from different os's. In general, the display has been abstracted, and now I'm working on the actual math.

+++ src/display/point.java
@@ -19,8 +19,8 @@ public class point implements Cloneable {
 			y = Y;
 		}
 		public point(float X, float Y ){
-			x = (int)X;
-			y = (int)Y;
+			x = X;
+			y = Y;
 			
 		}
